requirements.txt
pillow==10.2.0
pixelmatch==0.3.0
playwright==1.41.2
requests==2.31.0
urllib3==2.0.6
